Refactored Hud and all *system.update methods
Added stages for main menu, pause menu, and dead. These are yet to be implemented, but the skeleton is there.

Refactored *system.update() methods to not update if: JmpGame.gameState != GameState.Running. The exception to this is the HudSystem class. This still updates so that I can display Main menu, pause menu, and death screen when appropriate.